requirements.txt
Django==3.2.13
Pillow==9.1.0
coverage==6.3.2
stripe==2.72.0
django-mptt==0.13.4
django-countries==7.3.2
django-debug-toolbar==3.2.4
six==1.16.0
flake8==4.0.1
flake8-isort==4.1.1
black==22.3.0
geopy==2.2.0
geoip2==4.5.0
folium==0.12.1.post1
django-modeltranslation==0.18.2
celery==5.2.6
redis==4.3.1
django-celery-beat==2.2.1
django-celery-results==2.3.1
